Add AbstractMySQLRepository

+++ service/src/main/java/com/epam/esm/service/validator/TagValidator.java
@@ -35,7 +35,7 @@ public class TagValidator {
         }
     }
 
-    public static boolean idValidation(int id) {
+    public static boolean idValidation(long id) {
         boolean res = true;
 
         if (id < 0 || id > MAX_ID) {
